Rename type field in ModelType to model_type to avoid confusion, moved slurm template models to client_vars as they are not meant to be public facing, updated _models to be public
